Display init of Pixexl only one time
