+++ views/templates/hook/header.tpl
@@ -43,7 +43,6 @@
 {/strip}
 {literal}
     fbq('track', 'PageView');
-    var pixel_fc = {/literal}"{$pixel_fc|escape:'htmlall':'UTF-8'}"{literal};
 </script>
 
 <noscript>
@@ -82,5 +81,3 @@
   {/if}
 {/if}
 <!-- END OF Set Facebook Pixel Product Export -->
-
-{include file="./fbTrack.tpl"}
